flask/requirements.txt
python-dotenv==0.14.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Script==2.0.6
Flask-Migrate==2.5.3
Flask-Security==3.0.0
email-validator==1.1.1
uwsgi==2.0.19.1